graphing/requirements.txt
Package    Version
---------- -------
numpy      1.18.1
pip        20.3.1
PyQt5      5.14.1
PyQt5-sip  12.7.0
pyqtgraph  0.10.0
pyserial   3.5
setuptools 51.0.0
shiboken2  5.14.0
wheel      0.36.1
